Add files via upload
The software part mainly consists of 2 functions: clockwise and anticlockwise which are activated by a flag which is triggered by the SW1 button press. These functions are sequentially activating the pins in order to energise each coil.

In order for the motor to have more torque , we use both coils , a

technique called half stepping.  (using two “1” ’s instead of just one 1 per step . )

The motor will rotate clockwise, as a natural state. When we press the SW1 button it will rotate anticlockwise, then when SW2 is pressed the motor will turn clockwise again and so on. Basically the right switch is turning the motor to the right and the left switch is turning it to the left.